+++ build.py
@@ -42,26 +42,30 @@ def compile_openssl(openssl_version, script_steps, configure_args, distribution)
         # OpenSSL 1.0.x is problematic with concurrently builds so set the max to just 1.
         jobs = '1'
 
-    compile_openssl = '''wget \
-    -q -O '/tmp/openssl-{0}.tar.gz' \
+    openssl_configure_args = ['--prefix="{0}"'.format(build_path)]
+    if openssl_version.startswith('3.'):
+        openssl_configure_args.append('--libdir="{0}/lib"'.format(build_path))
+
+    openssl_configure_script = build_multiline_command(compile_arg, openssl_configure_args)
+
+    compile_openssl = '''wget \\
+    -q -O '/tmp/openssl-{0}.tar.gz' \\
     'https://www.openssl.org/source/openssl-{0}.tar.gz'
 
 tar -xf '/tmp/openssl-{0}.tar.gz' -C /tmp
 cd '/tmp/openssl-{0}'
 
-{1} \
-    '--prefix={2}'
+{1}
 make -j{3}
-make install_sw'''.format(openssl_version, compile_arg, build_path, jobs)
+make install_sw'''.format(openssl_version, openssl_configure_script, build_path, jobs)
 
     script_steps.append(('Compiling OpenSSL %s' % openssl_version, compile_openssl))
 
-    # TODO: Enable this once AZP opens up the Big Sur agents so we can actually run this in CI.
-    if distribution.startswith('macOS') and False:
+    if distribution.startswith('macOS'):
         # We want to create a fat (x64 and arm) library so we can compile mi for arm.
         compile_openssl = '''
-MACOSX_DEPLOYMENT_TARGET=10.15 ./Configure \
-    darwin64-arm64-cc shared \
+./Configure \\
+    darwin64-arm64-cc shared \\
     '--prefix={0}-arm64'
 make clean
 make -j
@@ -80,9 +84,9 @@ for file in "${{LIB_DIR}}"/lib/lib*; do
     fi
 done
 
-lipo -create \
-    '{0}/bin/openssl' \
-    '{0}-arm64/bin/openssl' \
+lipo -create \\
+    '{0}/bin/openssl' \\
+    '{0}-arm64/bin/openssl' \\
     -output '{0}/bin/openssl'
 '''.format(build_path)
         script_steps.append(('Compiling OpenSSL for arm64', compile_openssl))
@@ -190,7 +194,12 @@ cd repo'''))
     script_steps.append(('Applying psl-omi-provider patches', '\n'.join(['''echo "Applying '{0}'"
 git apply "${{OMI_REPO}}/psl-omi-provider/{0}"'''.format(p) for p in psl_patches])))
 
-    built_type = 'Debug' if args.debug else 'Release'
+    cmake_args = {'CMAKE_BUILD_TYPE': 'Debug' if args.debug else 'Release'}
+    if distribution.startswith('macOS'):
+        cmake_args['CMAKE_OSX_ARCHITECTURES'] = 'arm64;x86_64'
+
+    cmake_arg_string = " ".join(["-D%s='%s'" % (a, v) for a, v in cmake_args.items()])
+
     script_steps.append(('Building libpsrpclient', '''rm -rf omi
 ln -s "${{OMI_REPO}}" omi
 
@@ -201,9 +210,9 @@ ln -s {0} omi/Unix/output
 
 cd src
 echo -e "Running cmake with\\n\\t-DCMAKE_BUILD_TYPE={1}"
-cmake -DCMAKE_BUILD_TYPE={1} .
+cmake {1} .
 make psrpclient
-cp libpsrpclient.* "${{OMI_REPO}}/PSWSMan/lib/{2}/"'''.format(output_dirname, built_type, distribution)))
+cp libpsrpclient.* "${{OMI_REPO}}/PSWSMan/lib/{2}/"'''.format(output_dirname, cmake_arg_string, distribution)))
 
     if distribution.startswith('macOS'):
         script_steps.append(('Patch libmi dylib path for libpsrpclient',
@@ -231,15 +240,27 @@ for file in "${{LIB_DIR}}"/lib/lib*.dylib; do
             "${{file}}" \\
             "@loader_path/${{FILENAME}}" \\
             "${{OMI_REPO}}/PSWSMan/lib/{1}/libmi.dylib"
+
+        ARM64_LIB="$( dirname $( dirname "${{file}}" ) )-arm64/lib/${{FILENAME}}"
+        install_name_tool -change \\
+            "${{ARM64_LIB}}" \\
+            "@loader_path/${{FILENAME}}" \\
+            "${{OMI_REPO}}/PSWSMan/lib/{1}/libmi.dylib"
     fi
 done'''.format(openssl_version, distribution)))
 
         script_steps.append(('Output linked information',
             '''echo "libpsrpclient links"
-otool -L "${{OMI_REPO}}/PSWSMan/lib/{0}/libpsrpclient.dylib"
+otool -L -arch all "${{OMI_REPO}}/PSWSMan/lib/{0}/libpsrpclient.dylib"
 
 echo "libmi links"
-otool -L "${{OMI_REPO}}/PSWSMan/lib/{0}/libmi.dylib"
+otool -L -arch all "${{OMI_REPO}}/PSWSMan/lib/{0}/libmi.dylib"
+
+echo "libmi architecture info"
+lipo -archs "${{OMI_REPO}}/PSWSMan/lib/{0}/libmi.dylib"
+
+echo "libpsrpclient architecture info"
+lipo -archs "${{OMI_REPO}}/PSWSMan/lib/{0}/libpsrpclient.dylib"
 '''.format(distribution)))
 
     else: