+++ CHANGELOG.md
@@ -5,6 +5,7 @@ It documents the changes in each of the tagged releases
 
 ## 2.3.0 - TBD
 
++ Added universal build for macOS to work with both x86_64 and arm64 processes
 + Fixed up logic used to determine what OpenSSL library is used on macOS
 
 ## 2.2.1 - 2021-07-14